#5 修改问卷小数

See merge request edu/aizhixin_api_cloud!138

+++ diandian/src/main/java/com/aizhixin/cloud/dd/questionnaire/dto/QuestionnaireDTO.java
@@ -30,7 +30,7 @@ public class QuestionnaireDTO implements Serializable {
 
     @NotNull
     @ApiModelProperty(value = "问卷总分", required = true)
-    protected Integer totalScore;
+    protected Float totalScore;
 
     @ApiModelProperty(value = "创建时间", required = false)
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss", timezone = "GMT+8")
